requirements.txt
djangorestframework_api_key==2.0.0
drf_yasg==1.17.1
Django==3.0.8
djangorestframework==3.11.0
django-cors-headers==3.4.0
